refactor: type salaryWorkTimeStatistics selector and extract hooks to TypeScript
