+++ src/selectors/companyAndJobTitle.ts
@@ -7,6 +7,7 @@ import { RatingStatistics } from 'apis/queryCompanyRatingStatistics';
 import { CompanySalaryWorkTimeStatistics } from 'apis/queryCompanySalaryWorkTimeStatistics';
 import { TopNJobTitles } from 'apis/queryCompanyTopNJobTitles';
 import { JobTitleSalaryWorkTimeStatistics } from 'apis/queryJobTitleSalaryWorkTimeStatistics';
+import { SalaryWorkTimeStatistics } from 'apis/salaryWorkTime';
 import { RootState } from 'reducers';
 import {
   CompanyAspectExperienceResult,
@@ -27,9 +28,19 @@ import {
 } from 'reducers/jobTitleIndex';
 import FetchBox, { getUnfetched, isFetched } from 'utils/fetchBox';
 
+type SalaryWorkTimeStats = Pick<
+  SalaryWorkTimeStatistics,
+  | 'count'
+  | 'is_overtime_salary_legal_count'
+  | 'has_compensatory_dayoff_count'
+  | 'has_overtime_salary_count'
+>;
+
 export const salaryWorkTimeStatistics: (
-  box: FetchBox<unknown>,
-) => unknown = R.pipe(
+  box: FetchBox<
+    CompanySalaryWorkTimeStatistics | JobTitleSalaryWorkTimeStatistics | null
+  >,
+) => SalaryWorkTimeStats | Record<string, never> = R.pipe(
   state => state.data,
   R.when(R.is(Object), R.prop('salary_work_time_statistics')),
   R.defaultTo({}),
